refactor: split account identity edit workflow
