+++ frontend/src/accounts/detail/AccountDetailPage.tsx
@@ -8,7 +8,7 @@ import CreateTransactionModal from '@/components/CreateTransactionModal'
 import AccountIdentityCard from '@/accounts/detail/components/identity/AccountIdentityCard'
 import BackLink from '@/accounts/detail/components/BackLink'
 import BalanceChartCard from '@/accounts/detail/components/balance-chart/BalanceChartCard'
-import EditAccountIdentityModal from '@/accounts/detail/components/EditAccountIdentityModal'
+import EditAccountIdentityModal from '@/accounts/detail/components/edit-identity/EditAccountIdentityModal'
 import MonthlyCashFlowCard from '@/accounts/detail/components/monthly-cash-flow/MonthlyCashFlowCard'
 import { TopCategoriesBySpendingCard } from '@/accounts/detail/components/spending-breakdown/TopCategoriesBySpendingCard'
 import { TopMerchantsBySpendingCard } from '@/accounts/detail/components/spending-breakdown/TopMerchantsBySpendingCard'
