+++ frontend/src/accounts/detail/components/edit-identity/types.ts
@@ -1,0 +1,8 @@
+import type { IdentityFormValues } from '@/accounts/detail/utils/identityForm'
+
+export type DeleteStage = 'idle' | 'confirm' | 'type-name'
+
+export type SetIdentityFormField = <K extends keyof IdentityFormValues>(
+  field: K,
+  value: IdentityFormValues[K],
+) => void
